Moved the profile model made changes to the model
